refactor: Split exceptions into a package
